Code Refactored for Readability
Separated the save and import functionalities into separate functions.
